requirements.txt
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3

boto3==1.15.5
email-validator==1.1.1
pandas==1.1.2
requests==2.24.0
Pillow==8.1.0
pyotp==2.4.1
qrcode==6.1
pyOpenSSL==20.0.1
psycopg2-binary==2.8.6

braintree==4.5.0